Create README.md
For this project I scraped the top 10,000 grossing films from IMDb using BeautifulSoup and Python.
Then I cleaned the data thoroughly and did some feature engineering.
My main goal was to build a model to predict total Domestic Gross only using information that could be known before a movie is released in theaters.
I discovered that a director's mean gross of previous films was a strong positive predictor of box office success. Additionally,  runtime positively predicted box office gross as did a G, PG, or PG-13 rating (as compared to an R rating or being Not Rated). Other positive predictors of box office gross were number of films a director had in the top 10,000 and having a colon in the title.